Implement pp6calculator Day 3 Exercise 5 solution
Change FourVector from a `struct` to a `class` and move `interval`,
`boost_z` `getCausalType` and `asString` free functions to member
functions of the class.

Retain create/destroy free functions to keep menu interface as similar
as possible. Keep asString for `CausalType` (now
`FourVector::CausalType`) enum as a free function as it is not truly
part of the FourVector interface. Change calculations to message member
functions instead of free functions.