Implement pp6calculator Day 3 Exercise 7 solution
Make all data members and constants of FourVector private. Provide
getters/setters for the data members. Implement copy constructor in
terms of these getters/setters.

Cache the interval value as a data member of FourVector and recache it
everytime one of the FourVector components is changed. As we expect the
interval to be accessed more frequently than the data members are
changed, this should provide a small performance boost.