Implement pp6calculator Day 4 Exercise 1 solution
Add template method `getField` method to `FileReader` reading and
returning the supplied field as the templated type. If input fails,
set the failed flag to true and return a default constructed instance
of the template type.

Add supplied `pdg.dbt` to top level of project, and copy it to base
of project build directory so it's easily accessible to pp6calculator.

As per exercise, add operation to Day 4 menu to read in a .dbt format
file, iterate over the lines and print each field to stdout.