Implement pp6calculator Day 4 Exercise 5 solution
Add operation to Day 4 menu calling new `pp6day4_algorithm_demo`
function. Query user for size of vector to generate, restricting the
size to a number between 10 and 100. If the user enters an invalid
number, simply ask them again.

Create a vector with the required capacity, use `std::generate` to fill
the vector with random numbers from the PP6Lib getRandom function. Use
a `std::ostream_iterator` to report the contents before and then after
sorting it with the STL `sort` function. Also use the `std::max_element`
and `std::min_element` to find the maximum and minimum values if the
vector, reporting these to stdout.