Update GNU/Clang compile flag setting
Update check of compiler ID to apply to GNU and Clang compilers.
Remove -fno-nonansi-builtins from default set, only adding this
to list when GNU compiler is used (Clang does not recognise it).

Check for -stdlib with libc++ flag and apply it can be used and
we are on Apple (libc++ on platforms other than OS X and FreeBSD
is incomplete).

Check for availability of 0x and 11 C++ standards, using best
available and otherwise defaulting to C++98.